+++ src/error.rs
@@ -1,65 +1,58 @@
 //! All error types used through the library.
 
 use chrono::{DateTime, Utc};
-use failure::{Context, Fail};
+use thiserror::Error;
 
 /// Errors that can occur while creating or using a `Celery` app.
-#[derive(Debug, Fail)]
+#[derive(Error, Debug)]
 pub enum CeleryError {
-    /// The queue you're attempting to use has not been defined.
-    #[fail(display = "CeleryError: unknown queue '{}'", _0)]
-    UnknownQueue(String),
-
     /// Raised when `Celery::consume_from` is given an empty array of queues.
-    #[fail(display = "CeleryError: at least one queue required to consume from")]
+    #[error("at least one queue required to consume from")]
     NoQueueToConsume,
 
     /// Forced shutdown.
-    #[fail(display = "CeleryError: forced shutdown")]
+    #[error("forced shutdown")]
     ForcedShutdown,
 
     /// Any other broker-level error that could happen when initializing or with an open
     /// connection.
-    #[fail(display = "CeleryError: broker error ({})", _0)]
-    BrokerError(#[fail(cause)] BrokerError),
+    #[error("broker error")]
+    BrokerError(#[from] BrokerError),
 
     /// Any other IO error that could occur.
-    #[fail(display = "CeleryError: IO error ({})", _0)]
-    IoError(#[fail(cause)] std::io::Error),
+    #[error("IO error")]
+    IoError(#[from] std::io::Error),
 
     /// A protocol error.
-    #[fail(display = "CeleryError: protocol error ({})", _0)]
-    ProtocolError(#[fail(cause)] ProtocolError),
+    #[error("protocol error")]
+    ProtocolError(#[from] ProtocolError),
 
     /// An invalid glob pattern for a routing rule.
-    #[fail(display = "CeleryError: invalid glob routing rule '{}'", _0)]
-    BadRoutingPattern(#[fail(cause)] globset::Error),
+    #[error("invalid glob routing rule")]
+    BadRoutingPattern(#[from] globset::Error),
 
     /// There is already a task registerd to this name.
-    #[fail(
-        display = "CeleryError: there is already a task registered as '{}'",
-        _0
-    )]
+    #[error("there is already a task registered as '{0}'")]
     TaskRegistrationError(String),
 
-    #[fail(display = "CeleryError: received unregistered task {}", _0)]
+    #[error("received unregistered task {0}")]
     UnregisteredTaskError(String),
 }
 
 /// Errors that can occur while creating or using a `Beat` app.
-#[derive(Debug, Fail)]
+#[derive(Error, Debug)]
 pub enum BeatError {
     /// Any broker-level error.
-    #[fail(display = "BeatError: {}. Cause: {}", _0, _1)]
-    BrokerError(String, #[fail(cause)] BrokerError),
+    #[error("broker error")]
+    BrokerError(#[from] BrokerError),
 
     /// A protocol error.
-    #[fail(display = "BeatError: {}. Cause: {}", _0, _1)]
-    ProtocolError(String, #[fail(cause)] ProtocolError),
+    #[error("protocol error")]
+    ProtocolError(#[from] ProtocolError),
 }
 
 /// Errors that can occur at the task level.
-#[derive(Debug, Fail)]
+#[derive(Error, Debug)]
 pub enum TaskError {
     /// An error that is expected to happen every once in a while.
     ///
@@ -71,7 +64,7 @@ pub enum TaskError {
     /// If that service is temporarily unavailable the task should raise an `ExpectedError`.
     ///
     /// Tasks are always retried with capped exponential backoff.
-    #[fail(display = "TaskError (expected): {}", _0)]
+    #[error("task raised expected error: {0}")]
     ExpectedError(String),
 
     /// Should be used when a task encounters an error that is unexpected.
@@ -80,7 +73,7 @@ pub enum TaskError {
     /// when this error is encountered is determined by the
     /// [`TaskOptions::retry_for_unexpected`](../task/struct.TaskOptions.html#structfield.retry_for_unexpected)
     /// setting.
-    #[fail(display = "TaskError: {}", _0)]
+    #[error("task raised unexpected error: {0}")]
     UnexpectedError(String),
 
     /// Raised when a task runs over its time limit specified by the
@@ -92,7 +85,7 @@ pub enum TaskError {
     /// Typically a task implementation doesn't need to return these errors directly
     /// because they will be raised automatically when the task runs over it's `time_limit`,
     /// provided the task yields control at some point (like with non-blocking IO).
-    #[fail(display = "TaskError: task timed out")]
+    #[error("task timed out")]
     TimeoutError,
 
     /// A task can return this error variant to manually trigger a retry.
@@ -100,48 +93,48 @@ pub enum TaskError {
     /// This error variant should generally not be used directly. Instead, you should
     /// call the `Task::retry_with_countdown` or `Task::retry_with_eta` trait methods
     /// to manually trigger a retry from within a task.
-    #[fail(display = "TaskError: retry triggered")]
+    #[error("task retry triggered")]
     Retry(Option<DateTime<Utc>>),
 }
 
 /// Errors that can occur while tracing a task.
-#[derive(Debug, Fail)]
+#[derive(Error, Debug)]
 pub(crate) enum TraceError {
     /// Raised when a task throws an error while executing.
-    #[fail(display = "TraceError: task failed with {}", _0)]
+    #[error("task failed")]
     TaskError(TaskError),
 
     /// Raised when an expired task is received.
-    #[fail(display = "TraceError: task expired")]
+    #[error("task expired")]
     ExpirationError,
 
     /// Raised when a task should be retried.
-    #[fail(display = "retrying task")]
+    #[error("retrying task")]
     Retry(Option<DateTime<Utc>>),
 }
 
 /// Errors that can occur at the broker level.
-#[derive(Debug, Fail)]
+#[derive(Error, Debug)]
 pub enum BrokerError {
     /// Raised when a broker URL can't be parsed.
-    #[fail(display = "BrokerError: invalid broker URL '{}'", _0)]
+    #[error("invalid broker URL '{0}'")]
     InvalidBrokerUrl(String),
 
     /// The queue you're attempting to use has not been defined.
-    #[fail(display = "BrokerError: unknown queue '{}'", _0)]
+    #[error("unknown queue '{0}'")]
     UnknownQueue(String),
 
     /// Broker is disconnected.
-    #[fail(display = "BrokerError: broker not connected")]
+    #[error("broker not connected")]
     NotConnected,
 
     /// Any IO error that could occur.
-    #[fail(display = "BrokerError: IO error ({})", _0)]
-    IoError(#[fail(cause)] std::io::Error),
+    #[error("IO error")]
+    IoError(#[from] std::io::Error),
 
     /// Any other AMQP error that could happen.
-    #[fail(display = "BrokerError: AMQP error ({})", _0)]
-    AMQPError(#[fail(cause)] lapin::Error),
+    #[error("AMQP error")]
+    AMQPError(#[from] lapin::Error),
 }
 
 impl BrokerError {
@@ -160,87 +153,30 @@ impl BrokerError {
 }
 
 /// Errors that can occur due to messages not conforming to the protocol.
-#[derive(Debug, Fail)]
+#[derive(Error, Debug)]
 pub enum ProtocolError {
     /// Raised when a required message property is missing.
-    #[fail(display = "ProtocolError: missing required property '{}'", _0)]
+    #[error("missing required property '{0}'")]
     MissingRequiredProperty(String),
 
     /// Raised when the headers are missing altogether.
-    #[fail(display = "ProtocolError: missing headers")]
+    #[error("missing headers")]
     MissingHeaders,
 
     /// Raised when a required message header is missing.
-    #[fail(display = "ProtocolError: missing required property '{}'", _0)]
+    #[error("missing required property '{0}'")]
     MissingRequiredHeader(String),
 
     /// Raised when serializing or de-serializing a message body fails.
-    #[fail(display = "ProtocolError: serialization error ({})", _0)]
-    BodySerializationError(#[fail(cause)] serde_json::Error),
-}
-
-impl From<BrokerError> for CeleryError {
-    fn from(err: BrokerError) -> Self {
-        Self::BrokerError(err)
-    }
-}
-
-impl From<ProtocolError> for CeleryError {
-    fn from(err: ProtocolError) -> Self {
-        Self::ProtocolError(err)
-    }
-}
-
-impl From<std::io::Error> for CeleryError {
-    fn from(err: std::io::Error) -> Self {
-        Self::IoError(err)
-    }
-}
-
-impl From<globset::Error> for CeleryError {
-    fn from(err: globset::Error) -> Self {
-        Self::BadRoutingPattern(err)
-    }
-}
-
-impl From<serde_json::Error> for ProtocolError {
-    fn from(err: serde_json::Error) -> Self {
-        Self::BodySerializationError(err)
-    }
-}
-
-impl From<lapin::Error> for BrokerError {
-    fn from(err: lapin::Error) -> Self {
-        match err {
-            lapin::Error::IOError(e) => BrokerError::IoError(std::io::Error::new(
-                (*e).kind(),
-                format!("{} from AMQP broker", *e),
-            )),
-            lapin::Error::InvalidConnectionState(lapin::ConnectionState::Closing)
-            | lapin::Error::InvalidConnectionState(lapin::ConnectionState::Closed)
-            | lapin::Error::InvalidConnectionState(lapin::ConnectionState::Error)
-            | lapin::Error::InvalidChannelState(lapin::ChannelState::Closing)
-            | lapin::Error::InvalidChannelState(lapin::ChannelState::Closed)
-            | lapin::Error::InvalidChannelState(lapin::ChannelState::Error) => {
-                BrokerError::NotConnected
-            }
-            _ => BrokerError::AMQPError(err),
-        }
-    }
-}
-
-impl From<Context<&str>> for TaskError {
-    fn from(ctx: Context<&str>) -> Self {
-        Self::UnexpectedError((*ctx.get_context()).into())
-    }
+    #[error("serialization error")]
+    BodySerializationError(#[from] serde_json::Error),
 }
 
 /// Extension methods for `Result` types within a task body.
 ///
 /// These methods can be used to convert a `Result<T, E>` to a `Result<T, TaskError>` with the
 /// appropriate `TaskError` variant. The trait has a blanket implementation for any error type that implements
-/// [`std::error::Error`](https://doc.rust-lang.org/std/error/trait.Error.html) or
-/// [`failure::Fail`](https://docs.rs/failure/0.1.6/failure/trait.Fail.html).
+/// [`std::error::Error`](https://doc.rust-lang.org/std/error/trait.Error.html).
 pub trait TaskResultExt<T, E> {
     /// Convert the error type to a `TaskError::ExpectedError`.
     fn with_expected_err(self, context: &str) -> Result<T, TaskError>;
@@ -251,7 +187,7 @@ pub trait TaskResultExt<T, E> {
 
 impl<T, E> TaskResultExt<T, E> for Result<T, E>
 where
-    E: Fail,
+    E: std::error::Error,
 {
     fn with_expected_err(self, context: &str) -> Result<T, TaskError> {
         self.map_err(|_failure| TaskError::ExpectedError(context.into()))